Make glyph deletion work with DirectWrite
It does not seem to support glyph deletion via multiple substitutions
with empty glyph (AKA `sub foo by NULL`) in Arabic. So instead we now
delete the glyphs by substituting them with a dummy mark first, then
delete the mark in separate step. The dummy mark is ignored in most
lookups, so it has the same effect, but we also delete it so that in
engines that support it we keep the old behaviour.